refactor(common): 修复类型转换和导入缺失问题
- 修复 Integer 到 Long 类型转换问题，在分页结果返回时使用 longValue() 方法
- 修复 Result.success() 方法调用，统一添加 null 参数以符合返回格式
- 修复 RefundRequest 类缺少 BigDecimal 导入问题
- 修复 PaymentMapper 缺少 List 导入问题
- 修复 TaskLogService 缺少 Map 和 HashMap 导入问题
- 移除 xxl-job 相关依赖，改为本地日志处理方式
- 修复 Mongo 查询排序语法问题
- 移除 TaskLogRepository 中重复的方法定义
- 添加进度汇总文档，记录已完成工作和后续建议

+++ admin-service/src/main/java/com/learning/admin/service/impl/AuditLogServiceImpl.java
@@ -65,7 +65,7 @@ public class AuditLogServiceImpl implements AuditLogService {
         List<AuditLogDO> logs = mongoTemplate.find(query, AuditLogDO.class);
         long total = mongoTemplate.count(countQuery, AuditLogDO.class);
         List<AuditLogDTO> records = logs.stream().map(this::toDTO).toList();
-        return PageResult.of(records, total, request.getCurrent(), request.getSize());
+        return PageResult.of(records, total, request.getCurrent().longValue(), request.getSize().longValue());
     }
 
     private AuditLogDTO toDTO(AuditLogDO log) {
